Fixed a CSS toolchain issue.
`@import`-ing CSS would cause karma to blow up.

Need to make sure postcss is intialized there too.

Connected to rangle/rangle-starter#127
